requirements.txt
dash==2.2.0
dash-bootstrap-components==1.0.3
dash-daq==0.5.0
dash-renderer==1.9.1
dash-table==5.0.0
icalendar==4.0.9
numpy==1.22.3
pandas==1.4.1
pre-commit==2.17.0
python-dotenv==0.19.2
